Add a config option that always displays the mandatory fields
Before this change the fields that were flagged by the extension config
would only show on the display page when it is required that they be
givne a value. This update provides a configuration option to always
show the fields on the update page even if they are not required to be set
for the current update.