Restore filtering to show only mandatory fields
When EditCustomFields was moved from Tickets/Elements to
the general /Elements, the MassageCustomFields callback
was not also moved, so the code that filters the list of
CFs to only mandatory fields was lost. This resulted in
all CFs showing up on update pages. Restore this functionality
by creating the appropriate new callback at
BeforeCustomFields.

Resolves: https://rt.cpan.org/Ticket/Display.html?id=102746