Core RT trims whitespace before updating; do so for "emptyness" check
RT trims leading and trailing whitespace of CF values submitted before
checking them for emptyness or applying them.  Not doing so in
MandatoryOnTransition can allow one to skate past a mandatory CF by
entering spaces, which MandatoryOnTransition will treat as "non-empty"
but core RT will ignore.

Use _NormalizeObjectCustomFieldValue to trim whitespace as core RT
does before validating "emptyness" of CFs.  This also deals with
multiple-entry and multi-line CFs -- while exposing that
MandatoryOnTransition's behavior for them is not well-defined.
Specifically, it always checks the first value submitted or the first
existing value, which may be insufficient for `must_be` or
`must_not_be` checks.  This commit does not attempt to resolve that;
it does document this limitation, however.