Add support for Owner as a mandatory field
This makes it possible to require Owner to be set to anything but Nobody when
performing a transition. Typically, this would be used to require an owner to be
set before resolving a ticket.

Also added a test case for the new functionality.

Code mostly lifted from, with tweakings and the test case added:
 https://forum.bestpractical.com/t/prevent-resolution-of-ticket-that-owner-is-nobody/30804/5

Signed-off-by: Gergely Nagy <algernon@bestpractical.com>