Allow rules that apply only on ticket creation
MandatoryOnTransition supports '* -> foo' rules which can
apply on ticket creation. However, the '*' as the 'from' status means
that the rule would ALSO apply to ALL transitions that end up in the 'foo'
status.

A new flag value of '__CREATE__' is added to allow for the configuration
of rules that apply ONLY on ticket creation. The config for the example
above would be modified to '__CREATE__ -> foo' to apply only on
creation.